+++ tests/test_skoolhtml.py
@@ -1387,6 +1387,10 @@ class SkoolMacroTest(HtmlWriterTestCase, CommonSkoolMacroTest):
         # Method used to test the #CALL macro
         return str((cwd, arg1, arg2, arg3))
 
+    def _test_call_with_kwargs(self, cwd, arg1, arg2=None, arg3=None):
+        # Method used to test the #CALL macro with keyword arguments
+        return str((arg1, arg2, arg3))
+
     def _test_call_no_retval(self, *args):
         return